add: distributions/general family (*lss)
